Moved `PanGestureRecognizer` proxy to it’s own file.
