+++ Sources/DefaultAnimator.swift
@@ -307,18 +307,3 @@ extension DefaultAnimator {
 		}
 	}
 }
-
-final class PanGestureRecognizer : UIPanGestureRecognizer {
-
-	var action: ( /* @escaping */ (UIPanGestureRecognizer) -> Void)?
-
-	init() {
-		super.init(target: nil, action: nil)
-		self.maximumNumberOfTouches = 1
-		self.addTarget(self, action: #selector(self.handlePan(_:)))
-	}
-
-	@objc private func handlePan(_ gesture: UIPanGestureRecognizer) {
-		self.action?(gesture)
-	}
-}
